refactor: Amélioration de la présentation de la section Blockchain
- Suppression des données relatives aux agences bancaires
- Réduction de la taille des graphiques
- Réorganisation de la barre de navigation
- Mise à jour du CHANGELOG
